Revert "Switch fpp_t and f_cnt_t to std::uint64_t"
This reverts commit e01046fc9dd67b535f60eb40edbccdc5dc837b4c.

+++ include/LmmsTypes.h
@@ -25,6 +25,7 @@
 #ifndef LMMS_TYPES_H
 #define LMMS_TYPES_H
 
+#include <cstddef>
 #include <cstdint>
 
 
@@ -40,8 +41,8 @@ using sample_t      = float;         // standard sample-type
 using int_sample_t  = std::int16_t;  // 16-bit-int-sample
 
 using sample_rate_t = std::uint32_t; // sample-rate
-using fpp_t         = std::uint64_t; // frames per period (0-16384)
-using f_cnt_t       = std::uint64_t; // standard frame-count
+using fpp_t         = std::size_t;   // frames per period (0-16384)
+using f_cnt_t       = std::size_t;   // standard frame-count
 using ch_cnt_t      = std::uint8_t;  // audio channel index/count (0-MaxChannelsPerAudioBuffer)
 using bpm_t         = std::uint16_t; // tempo (MIN_BPM to MAX_BPM)
 using bitrate_t     = std::uint16_t; // bitrate in kbps
